# doc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Circular Democracy API",
    "description": "API for processing citizen messages to politicians"
  },
  "servers": [
    {
      "url": "https://api.circulardemocracy.org",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8787",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/worker/process-replies": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Worker"
        ],
        "summary": "/api/v1/worker/process-replies",
        "description": "Manually trigger scheduled reply processing for admin use",
        "responses": {
          "200": {
            "description": "Scheduled replies processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "sent": {
                          "type": "number"
                        },
                        "failed": {
                          "type": "number"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "message_id": {
                                "type": "number"
                              },
                              "error": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "message_id",
                              "error"
                            ]
                          }
                        }
                      },
                      "required": [
                        "total",
                        "sent",
                        "failed",
                        "errors"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "result"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Worker processing failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}